Import of release hp2xx-3.4.3_a07.tar.gz 3.4.3a7: Stray Escape chars in the HPGL code are reported and ignored (one guy was trying to use esc-prefixed commands in his code because he was too familiar with PCL). Guess what, there was still some brokenness in IW/SC handling.:-( Fixed some wreckage from the penwidth type change in to_emf.c and added the ability to create EMF files on UNIX with libEMF. Added VS and CR support. Added -e for extra margin around IW (at least one commercial code miscalculates its clip boxes, cutting of text from spectra). FP with --nofill now does EP, so fat lettering done with just FP does not simply vanish.

+++ sources/picbuf.c
@@ -79,7 +79,7 @@ swapout_RowBuf (RowBuf *row, const PicBuf *picbuf)
 	exit	(ERROR);
   }
 
-  if (fwrite((char *) row->buf, picbuf->nb, picbuf->depth, picbuf->sd)
+  if ((int)fwrite((char *) row->buf, picbuf->nb, picbuf->depth, picbuf->sd)
 	!= picbuf->depth)
   {
 	PError	("swapout_RowBuf (on write)");
@@ -98,7 +98,7 @@ swapin_RowBuf (RowBuf *row, const PicBuf *picbuf)
 	exit	(ERROR);
   }
 
-  if (fread ((char *) row->buf, picbuf->nb, picbuf->depth, picbuf->sd)
+  if ((int)fread ((char *) row->buf, picbuf->nb, picbuf->depth, picbuf->sd)
 	!= picbuf->depth)
   {
 	PError	("swapin_RowBuf (on read)");
@@ -474,7 +474,7 @@ int	nr, not_allocated;
  **/
 
 	for (nr=0; nr < pb->nr; nr++)
-	    if (fwrite((char *) pb->row[0].buf, pb->nb, pb->depth, pb->sd)
+	    if ((int)fwrite((char *) pb->row[0].buf, pb->nb, pb->depth, pb->sd)
 		!= pb->depth)
 	    {
 			Eprintf ("Couldn't clear swap file!\n");
